refactor(createRoom): move DTOs about `Room` into `dto/Room` (#34)
- 프로젝트 관리를 위해 디렉토리 구조의 가독성을 높이기 위함입니다.
